HADOOP-15625 Patch 007
* added fs.s3.change.detection.versionrequired config
* fixed failure to throw RemoteFileChangedException on multiple reads
* added config property documentation to index.md

The fix to throw RemoteFileChangedException on multiple reads currently
means the 'warn' setting would result in potentially lots of warnings.
An S3AInputStream that detects a change would log a warn on every
subsequent read() call, which would be noisy, at least within the job or
task reading that file.  It probably does need to be revisited.